+++ Changelog.md
@@ -8,6 +8,9 @@
 - Show route start and end markers on the map ([issue #31](https://github.com/sandermvanvliet/RoadCaptain/issues/31))
 - Show KOM start (red) and finish (green) markers on the map ([issue #31](https://github.com/sandermvanvliet/RoadCaptain/issues/31)):
 ![Screenshot of Route Builder map with KOM markers shown in green and red](./images/route-builder-koms.png)
+- Show sprint start (red), finish (green) and segment (purple) on the map ([issue #31](https://github.com/sandermvanvliet/RoadCaptain/issues/31)):
+![Screenshot of Route Builder map with sprints markers shown in green and red](./images/route-builder-sprints.png)
+
 
 ## 0.6.0.0